passport_ocr/requirements.txt
opencv-python>=4.8.0
pytesseract>=0.3.10
paddlepaddle>=2.5.0
paddleocr>=2.7.0
numpy>=1.24.0,<2.0.0
pdf2image>=1.16.3
Pillow>=10.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
redis>=5.0.0
python-dotenv>=1.0.0

